# source/agri-chain/data/http-requests/agri_chain-openapi.yaml
openapi: "3.1.0"
info:
  title: "agri_chain API"
  description: "agri_chain API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
...
  /admin/all:
    get:
      summary: "GET admin/all"
      operationId: "getAllUsers"
      parameters:
        - name: "user"
          in: "query"
          required: true
          schema:
            $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/User"
  /admin/{id}/{status}:
    get:
      summary: "GET admin/{id}/{status}"
      operationId: "getUserById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "status"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "PENDING"
              - "ACTIVE"
              - "INACTIVE"
              - "BANNED"
              - "DELETED"
        - name: "user"
          in: "query"
          required: true
          schema:
            $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /admin/{id}:
    get:
      summary: "GET admin/{id}"
      operationId: "getUserById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "user"
          in: "query"
          required: true
          schema:
            $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    Timestamp:
      type: "object"
      properties:
        nanos:
          type: "integer"
          format: "int32"
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        role:
          type: "string"
          enum:
            - "GUEST"
            - "CUSTOMER"
            - "PRODUCER"
            - "TRANSFORMER"
            - "DISTRIBUTOR"
            - "EVENT_PLANNER"
            - "MODERATOR"
            - "CUSTOMER_SERVICE"
            - "TIME"
            - "ADMIN"
        username:
          type: "string"
        password:
          type: "string"
        birthDate:
          $ref: "#/components/schemas/Timestamp"
...